+++ system/src/sheets/ItemSheetSD.mjs
@@ -255,6 +255,7 @@ export default class ItemSheetSD extends ItemSheet {
 				"Ancestry",
 				"Armor",
 				"Basic",
+				"Boon",
 				"Class Ability",
 				"Class",
 				"Deity",
@@ -309,12 +310,29 @@ export default class ItemSheetSD extends ItemSheet {
 			await this.getClassSelectorConfigs(context);
 		}
 
-		if (["Spell"].includes(item.type)) {
+		if (["Scroll", "Spell", "Wand"].includes(item.type)) {
 			await this.getSpellSelectorConfigs(context);
 		}
 
 		if (["Armor", "Weapon"].includes(item.type)) {
 			context.propertyItems = await item.propertyItems();
+
+			const mySlug = item.name.slugify();
+
+			if (item.type === "Armor") {
+				context.baseArmor = await shadowdark.utils.getSlugifiedItemList(
+					await shadowdark.compendiums.baseArmor()
+				);
+
+				delete context.baseArmor[mySlug];
+			}
+			if (item.type === "Weapon") {
+				context.baseWeapons = await shadowdark.utils.getSlugifiedItemList(
+					await shadowdark.compendiums.baseWeapons()
+				);
+
+				delete context.baseWeapons[mySlug];
+			}
 		}
 
 		if (item.type === "Basic" && item.system.light.isSource) {
@@ -350,18 +368,6 @@ export default class ItemSheetSD extends ItemSheet {
 		if (item.type === "Talent" || item.type === "Effect" || item.system.magicItem ) {
 			context.predefinedEffects = await this._getPredefinedEffectsList();
 			context.effects = item.effects;
-		}
-
-		if (["Potion", "Scroll", "Spell", "Wand"].includes(item.type)) {
-			context.casterClasses = [];
-
-			for (const key of this.item.system.class) {
-				context.casterClasses.push(
-					CONFIG.SHADOWDARK.SPELL_CASTER_CLASSES[key]
-				);
-			}
-
-			context.casterClassesDisplay = context.casterClasses.join(", ");
 		}
 
 		context.descriptionHTML = await TextEditor.enrichHTML(
@@ -942,11 +948,15 @@ export default class ItemSheetSD extends ItemSheet {
 	 */
 	async _createPredefinedEffect(key, data) {
 		const handledData = data;
-		handledData.defaultValue = await this.item._handlePredefinedEffect(key, data.defaultValue);
 
-		if (handledData.defaultValue === "REPLACEME") {
+		let defaultValue = "REPLACEME";
+		[defaultValue] = await this.item._handlePredefinedEffect(key, data.defaultValue);
+
+		if (defaultValue === "REPLACEME") {
 			return shadowdark.log("Can't create effect without selecting a value.");
 		}
+
+		handledData.defaultValue = defaultValue;
 
 		const effectMode = foundry.utils.getProperty(
 			CONST.ACTIVE_EFFECT_MODES,