+++ RELEASE_NOTES.md
@@ -1,3 +1,14 @@
+## v.1.4.5
+
+### Bugfixes
+* [#515] Backstab option not available on attack rolls
+* [#516] Wand and Scroll item sheets not showing spell caster classes correctly
+
+### Enhancements
+* [#499] Add support for Boons and their various subtypes
+* [#517] Add support for limited use Class Abilities, and Class Abilities with no associated rolls
+* [#518] Weapon and Armor mastery using fixed base weapons rather than dynamically building a list from items available in compendiums
+
 ## v.1.4.4
 
 ### Bugfixes
@@ -30,7 +41,7 @@
 * [#493] Compendium item "Sword of the Ancients" missing base weapon type
 * [#496] Mage Armor spell effect not being applied
 
-### Enhancement
+### Enhancements
 * [#71] Add the ability to extend various character options
 * [#374] Added ability to create custom Classes
 * [#375] Added support for custom ancestries